+++ pkgs/universal_storage_sync/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-dev.3] - 2025-06-27
+
+- feat: Add `FilePathConfig` and `FilePath` models for storing file path and macOS bookmark.
+- feat: moved `MacOSBookmark` from `utils` package.
+
 ## [0.1.0-dev.2] - 2025-06-22
 
 - perf: Readme update